vsx color extended, warning fix

+++ lib/common/include/profiler/vsx_profiler_consumer.h
@@ -271,7 +271,7 @@ public:
 
       if (chunk.id != thread_id)
       {
-        vsx_printf(L"chunk id: %ld\n", chunk.id);
+        vsx_printf(L"chunk id: %ld\n", (long)chunk.id);
         continue;
       }
